# docs/swagger.yaml
basePath: /api/v1
...
host: https://api.7375.org
info:
  contact:
    email: info@7375.org
    name: Sergei
  description: API для управления видео-контентом BodyBalance.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: BodyBalance API
  version: "1.0"
paths:
...
  /img/{filename}:
    get:
      consumes:
      - application/json
      description: Send img file by filename
      parameters:
      - description: Video filename (e.g. 'shee_video.jpg')
        in: path
        name: filename
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: file
        "403":
          description: Forbidden
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Serve img file
      tags:
      - Files
...
  /video/{filename}:
    get:
      consumes:
      - application/json
      description: Streams video file by filename
      parameters:
      - description: Video filename (e.g. 'Sheya_baza.mp4')
        in: path
        name: filename
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: file
        "403":
          description: Forbidden
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Serve video file
      tags:
      - Files
...
swagger: "2.0"